Make OrderDetails more modular in InvoiceReport
Similar functionality is grouped, and the additional charge for the
colour red is moved out into a global constant. The caveat is that in
future, if a different extra charge based on shape is introduced, then
this would need to be refactored. However, for the current set of
requirements it is sufficient.